+++ app/Http/Controllers/Api/Partner/V1/CompanyController.php
@@ -4,12 +4,12 @@ namespace App\Http\Controllers\Api\Partner\V1;
 
 use App\Dtos\CompanyDto;
 use App\Http\Controllers\Controller;
-/*use App\JsonApi\Admin\V1\Companies\CompanyQuery;
-use App\JsonApi\Admin\V1\Companies\CompanyRequest;
-use App\JsonApi\Admin\V1\Companies\CompanySchema;
-use App\JsonApi\Admin\V1\CompanyUsers\CompanyUserCollectionQuery;*/
-//use App\Models\Partner\Company;
+use App\JsonApi\Partner\V1\Companies\CompanyQuery;
+use App\JsonApi\Partner\V1\Companies\CompanyRequest;
+use App\JsonApi\Partner\V1\Companies\CompanySchema;
+use App\Models\Company;
 use App\Services\CompanyService;
+use App\JsonApi\Proxies\CompanyPartner; // proxy model https://laraveljsonapi.io/docs/1.0/digging-deeper/proxies.html
 use LaravelJsonApi\Core\Responses\DataResponse;
 use LaravelJsonApi\Laravel\Http\Controllers\Actions;
 use LaravelJsonApi\Core\Document\Error;
@@ -19,7 +19,7 @@ class CompanyController extends Controller
 {
 
     use Actions\FetchMany;
-    use Actions\FetchOne;
+//    use Actions\FetchOne;
 //    use Actions\Store;
 //    use Actions\Update;
 //    use Actions\Destroy;
@@ -40,5 +40,29 @@ class CompanyController extends Controller
     public function __construct()
     {
         $this->companyService = new CompanyService();
+    }
+
+    /**
+     * Fetch zero to one JSON API resource by id.
+     *
+     * @param CompanySchema $schema
+     * @param CompanyQuery $request
+     * @param CompanyPartner $partner
+     * @return \Illuminate\Contracts\Support\Responsable|\Illuminate\Http\Response
+     */
+    public function show(CompanySchema $schema, CompanyQuery $request, CompanyPartner $partner)
+    {
+        //dd($partner);
+
+        $model = $schema
+            ->repository()
+            ->queryOne($partner)
+            ->withRequest($request)
+            ->first();
+
+
+        // do something custom...
+
+        return new DataResponse($model);
     }
 }